GraphQL Authorization

+++ packages/backend/src/modules/epubs/EPub.resolver.ts
@@ -1,4 +1,3 @@
-import { writeFileSync } from "fs";
 import { v4 as uuidv4 } from "uuid";
 import path from "path";
 import {
@@ -10,12 +9,13 @@ import {
   FieldResolver,
   Root,
   Ctx,
+  Authorized,
 } from "type-graphql";
 import { Repository } from "typeorm";
 import { InjectRepository } from "typeorm-typedi-extensions";
 
 import { ConnectionArguments } from "Relay/generics/ConnectionsArguments";
-import { connectionFromRepository } from "Relay/Connection.factory";
+import { connectionFromRelation, connectionFromRepository } from "Relay/Connection.factory";
 import { CreateEPubInput, UpdateEPubInput } from "Modules/epubs/EPub.inputs";
 import { UserInputError } from "apollo-server";
 
@@ -36,6 +36,7 @@ export default class EPubResolver {
   @InjectRepository(EPub, "prod")
   private readonly repository!: Repository<EPub>;
 
+  @Authorized(["ADMIN", "USER"])
   @FieldResolver(() => User, { nullable: true })
   async createdBy(@Root() epub: EPub): Promise<User> {
     return await this.repository
@@ -45,6 +46,7 @@ export default class EPubResolver {
       .loadOne();
   }
 
+  @Authorized(["ADMIN", "USER"])
   @FieldResolver(() => Book)
   async book(@Root() epub: EPub): Promise<Book> {
     return await this.repository
@@ -54,20 +56,35 @@ export default class EPubResolver {
       .loadOne();
   }
 
+  @Authorized("ADMIN")
   @Query(() => EPubConnection)
   async epubs(@Args() args: ConnectionArguments): Promise<EPubConnection> {
     return connectionFromRepository(args, this.repository);
   }
 
+  @Authorized(["ADMIN", "USER"])
+  @Query(() => EPubConnection)
+  async myEpubs(@Args() args: ConnectionArguments, @Ctx() { db, user }: Context): Promise<EPubConnection> {
+    return connectionFromRelation(args, db, User, "epubs", user);
+  }
+
+  @Authorized(["ADMIN", "USER"])
   @Query(() => EPub)
-  async epubById(@Arg("epubId") epubId: string): Promise<EPub> {
+  async epubById(@Arg("epubId") epubId: string, @Ctx() { user }: Context): Promise<EPub> {
     try {
-      return await this.repository.findOneOrFail(epubId);
+      const epub = await this.repository.findOneOrFail(epubId, { relations: ["createdBy"]});
+
+      if (epub.createdBy.id !== user.id) {
+        throw new UserInputError("No Permissions");
+      }
+
+      return await epub;
     } catch (error) {
       throw new UserInputError("No EPub Found with this id");
     }
   }
 
+  @Authorized(["ADMIN", "USER"])
   @Mutation(() => EPub)
   async uploadEPub(
     @Arg("data") { upload }: CreateEPubInput,
@@ -116,19 +133,32 @@ export default class EPubResolver {
     return epub;
   }
 
+  @Authorized(["ADMIN", "USER"])
   @Mutation(() => EPub)
   async updateEPub(
     @Arg("epubId") epubId: string,
-    @Arg("data") input: UpdateEPubInput
+    @Arg("data") input: UpdateEPubInput,
+    @Ctx() { user } : Context,
   ): Promise<EPub> {
-    await this.repository.update(epubId, input);
-    return this.repository.findOne(epubId);
+    const epub = await this.repository.findOneOrFail(epubId, { relations: ["createdBy"]});
+    
+    if (epub.createdBy.id !== user.id) {
+      throw new UserInputError("No Permissions");
+    }
+
+    return await this.repository.save(this.repository.merge(epub, input));
   }
 
+  @Authorized(["ADMIN", "USER"])
   @Mutation(() => EPub)
-  async deleteEPub(@Arg("epubId") epubId: string): Promise<EPub> {
+  async deleteEPub(@Arg("epubId") epubId: string, @Ctx() { user } : Context): Promise<EPub> {
     try {
-      const epub = await this.repository.findOneOrFail(epubId);
+      const epub = await this.repository.findOneOrFail(epubId, { relations: ["createdBy"]});
+    
+      if (epub.createdBy.id !== user.id) {
+        throw new UserInputError("No Permissions");
+      }
+
       await this.repository.remove(epub);
       return epub;
     } catch (error) {
